Agregado ver notas

+++ app/src/main/java/com/example/schoolcontrol/models/CreateGradeDto.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class CreateGradeDto(
     @SerializedName("student_id") val studentId: Int,
     @SerializedName("evaluation_id") val evaluationId: Int,
-    val score: Float
+    val score: Double
 )
